requirements.txt
flask==3.1.0
gunicorn==23.0.0
openai==1.82.0
PyPDF2==3.0.1
python-dotenv==1.1.0
werkzeug==3.1.3
fpdf2==2.8.7
